requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiopg==1.4.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cryptocompare==0.7.6
exceptiongroup==1.2.1
frozenlist==1.4.1
idna==3.7
importlib_metadata==7.1.0
loguru==0.7.2
magic-filter==1.0.12
multidict==6.0.5
packaging==24.0
peewee==3.17.5
peewee-async==0.10.0
peewee-migrate==1.12.2
psycopg2-binary==2.9.9
pycron==3.0.0
pydantic==2.7.1
pydantic_core==2.18.2
python-decouple==3.8
pytz==2024.1
redis==5.0.4
requests==2.32.2
sniffio==1.3.1
taskiq==0.11.3
taskiq-aiogram==0.3.1
taskiq-dependencies==1.5.3
taskiq-redis==0.5.6
typing_extensions==4.12.0
urllib3==2.2.1
yarl==1.9.4
zipp==3.19.1
